server/requirements.txt
astroid==1.6.4
autopep8==1.3.5
Django==2.0.5
django-cors-headers==2.2.0
django-filter==1.1.0
django-model-repr==0.0.6
django-query-parameters==0.2.3
djangorestframework==3.8.2
djangorestframework-camel-case==0.2.0
easy-thumbnails==2.5
gunicorn==19.8.1
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
Pillow==5.1.0
pycodestyle==2.4.0
pylint==1.9.1
pylint-django==0.11.1
pylint-plugin-utils==0.2.6
pytz==2018.4
redis==2.10.6
six==1.11.0
sorl-thumbnail==12.4.1
wrapt==1.10.11
